Got the keyword search
Rewrited from scratch the pov-keyword-help function and commented out
the old one, because the povuser.txt no more exists. Now the
pov-mode.el open an external browser to find the documentation. Try
this and let me know! Browser can be choosen in the M-x
customize-group pov menu. However, don't modify the name of directory
and file names if this feature works. You'll mess up things. This has
been tested on the documentation for pov-ray 3.61 (official binaries
for linux). Your documentation (for Mac or for previous or later
version of PoV-ray) may be somewhere else. Please let me know.