Fixed imenu bug
Cleaned the InsertMenu directory and added a pair of templates. If you
like this mode, please *donate* a spare scene to be inserted in the
InserMenu. Fixed the imenu critical bug that messed up the menubar.
Not tested with xemacs I consider this release quite stable. If you
find a bug feel free to contact me. I'll try to fix it.