Added a lot of keybindings and bug fixed
Fixed a minor bug that prevented the viewer to get the file name. This
happens when the path directory is too long and the output of povray
wrap words. I didn't discard the old code. If the buffer search fails,
it use the basename of current buffer appending the default extension
(png for unix, bmp for windoze). Added a lot of keybindings for
rendering and viewers.