Bug fixed and solved the InsertMenu issue
Dropped xemacs compatibility (I can't maintain this). Cleaned code.
The InsertMenu becomes just an add-on. Added the code to turn on and
 off the support to this feature.
Added the control to the Errors with exit code 0 (!)
Added the var "pov-isosurface-keyword" that was declared but void

2008-03-19
Added documentation (thanks to Xah Lee)
Solved the InsertMenu question with the POV-Ray team. Thanks to
 Christoph Hormann <http://www.imagico.de>, one of the authors of
 the InsertMenu. This package is covered by the POV-Ray license.
There is a commented function pov-online-search for keyword
 lookup in the on-line documentation, but still needs a lot of
 development.